Fixed initial temperature
When a ramp was used, the temperature passed was not being used
properly, since the initialize state function still required a
temperature argument to be passed. However, when using a ramp this is
cumbersome, so now an auxiliary function parses the information from the
ramp and uses the highest temperature to initialize the velocities. When
a single number is passed instead of a ramp, that is used instead.