Type stabilities and performance
In this case, some types got annotated with parametric types that have a
type hierarchy, following the performance advice from the manual.
Additionally there is some performance gains by having less allocations
now that I am using mutable MVectors. The difference between SVectors
and MVectors is almost null, but then it is easier to handle and
maintain, as it is more readable.