Fixes everything and performance gained
Now with these changes we are closer to LAMMPS, and the code is now 2.5x
slower, which is an improvement from being 6x slower. Neighbor lists are
used to loop over neighbors, but since the distance is not wrapped, I
had to wrap it myself. All parallelization is now turned off, and a
frequency of updating the lists is now implemented, which improves the
runtime by updating the lists much less often.